tokentofriend revamp
