+++ Tokens/tokentofriend/clients/build.gradle
@@ -38,7 +38,7 @@ dependencies {
 }
 
 springBoot {
-    mainClassName = "com.tokentofriend.webserver.Server"
+    mainClassName = "net.corda.samples.tokentofriend.webserver.Server"
 }
 
 task copyWebApp(type: Copy) {
@@ -78,7 +78,7 @@ node {
  */
 task runOperatorClient(type: JavaExec, dependsOn: assemble) {
     classpath = sourceSets.main.runtimeClasspath
-    main = 'com.tokentofriend.Client'
+    main = 'net.corda.samples.tokentofriend.Client'
     args 'localhost:10006', 'user1', 'test'
 }
 
@@ -87,7 +87,7 @@ task runOperatorClient(type: JavaExec, dependsOn: assemble) {
  */
 task runOperatoreServer(type: JavaExec, dependsOn: assemble) {
     classpath = sourceSets.main.runtimeClasspath
-    main = 'com.tokentofriend.webserver.Starter'
+    main = 'net.corda.samples.tokentofriend.webserver.Starter'
     args '--server.port=10050', '--config.rpc.host=localhost', '--config.rpc.port=10006', '--config.rpc.username=user1', '--config.rpc.password=test'
 }
 
@@ -96,6 +96,6 @@ task runOperatoreServer(type: JavaExec, dependsOn: assemble) {
  */
 task runUSWest1Server(type: JavaExec, dependsOn: assemble) {
     classpath = sourceSets.main.runtimeClasspath
-    main = 'com.tokentofriend.webserver.Starter'
+    main = 'net.corda.samples.tokentofriend.webserver.Starter'
     args '--server.port=10053', '--config.rpc.host=localhost', '--config.rpc.port=10015', '--config.rpc.username=user1', '--config.rpc.password=test'
 }
