--- Tokens/tokentofriend/clients/src/main/java/com/tokentofriend/webserver/Starter.java
@@ -1,23 +1,0 @@
-package com.tokentofriend.webserver;
-
-import org.springframework.boot.Banner;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-import static org.springframework.boot.WebApplicationType.SERVLET;
-
-/**
- * Our Spring Boot application.
- */
-@SpringBootApplication
-public class Starter {
-    /**
-     * Starts our Spring Boot application.
-     */
-    public static void main(String[] args) {
-        SpringApplication app = new SpringApplication(Starter.class);
-        app.setBannerMode(Banner.Mode.OFF);
-        app.setWebApplicationType(SERVLET);
-        app.run(args);
-    }
-}
