+++ Tokens/tokentofriend/gradle.properties
@@ -1,3 +1,3 @@
-name=Test
+name=Tokentofriend Cordapp
 group=com.tokentofriend
-version=0.1
+version=1.0
